separate styles for Signup
